Fixed a compile issue introduced with the last commit.

+++ src/_shared_/include/Devices/isx3/com_interface_codec.hpp
@@ -224,7 +224,7 @@ private:
   bool decodeImpedanceData(const std::vector<unsigned char> &payload,
                            short &fNumber, float &timestamp,
                            short &channelNumber,
-                           std::complex<float> &impedance);
+                           std::complex<double> &impedance);
 
   /**
    * @brief Decodes an acknowledgment payload received from the device.
